+++ .gitignore
@@ -7,3 +7,4 @@ npm-debug.log
 .env
 .db
 .idea
+dev.sqlite
