+++ routes/content.js
@@ -2,43 +2,12 @@ var express = require('express');
 var glob = require('glob');
 var fs = require('fs');
 var router = express.Router();
+const contentController = require('../controllers/contentController');
 
 //본문 조회
-router.get('/description/:page', async (req, res) => {
-	const contentId = req.params.page;
-	try {
-		fs.readFile('./public/content/description/' + contentId + '.md', 'utf8', (err, data) => {
-			if (err) {
-				return res.status(500).send({ error: 'not vaild content' });
-			} else {
-				let content = {};
-				content.title = contentId;
-				content.content = data;
-				res.send(content);
-			}
-		});
-	} catch {
-		res.status(500).send('Internal Server Error');
-	}
-});
+router.get('/description/:page', contentController.descriptionPage);
 
 //코드 조회
-router.get('/code/:page', async (req, res) => {
-	const contentId = req.params.page;
-	try {
-		fs.readFile('./public/content/python_problem/' + contentId + '.md', 'utf8', (err, data) => {
-			if (err) {
-				return res.status(500).send({ error: 'not vaild content' });
-			} else {
-				let content = {};
-				content.title = contentId;
-				content.content = data;
-				res.send(content);
-			}
-		});
-	} catch {
-		res.status(500).send('Internal Server Error');
-	}
-});
+router.get('/code/:page', contentController.codePage);
 
 module.exports = router;
